Fixed export of error bars.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@2393 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/plot2D/ErrorBarsCurve.h
@@ -49,8 +49,8 @@ public:
 	QwtArray<double> errors(){return err;};
 	void setErrors(const QwtArray<double>&data){err=data;};
 
-	int capLength(){return d_cap_length;};
-	void setCapLength(int t){d_cap_length = t;};
+	double capLength(){return d_cap_length;};
+	void setCapLength(double t){d_cap_length = t;};
 
 	double width(){return pen().widthF();};
 	void setWidth(double w);
@@ -101,7 +101,7 @@ private:
 	int type;
 
 	//! Length of the bar cap decoration
-	int d_cap_length;
+	double d_cap_length;
 
 	bool plus, minus, through;
